@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        v4.25.1
 // source: api/pb/products.proto
 
@@ -1029,8 +1029,8 @@ type ListCategoryItemsReq struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Page         string `protobuf:"bytes,1,opt,name=page,proto3" json:"page,omitempty"`
-	Size         string `protobuf:"bytes,2,opt,name=size,proto3" json:"size,omitempty"`
+	Page         uint32 `protobuf:"varint,1,opt,name=page,proto3" json:"page,omitempty"`
+	Size         uint32 `protobuf:"varint,2,opt,name=size,proto3" json:"size,omitempty"`
 	Sort         string `protobuf:"bytes,3,opt,name=sort,proto3" json:"sort,omitempty"`
 	CategorySlug string `protobuf:"bytes,4,opt,name=category_slug,json=categorySlug,proto3" json:"category_slug,omitempty"`
 }
@@ -1067,18 +1067,18 @@ func (*ListCategoryItemsReq) Descriptor() ([]byte, []int) {
 	return file_api_pb_products_proto_rawDescGZIP(), []int{12}
 }
 
-func (x *ListCategoryItemsReq) GetPage() string {
+func (x *ListCategoryItemsReq) GetPage() uint32 {
 	if x != nil {
 		return x.Page
 	}
-	return ""
+	return 0
 }
 
-func (x *ListCategoryItemsReq) GetSize() string {
+func (x *ListCategoryItemsReq) GetSize() uint32 {
 	if x != nil {
 		return x.Size
 	}
-	return ""
+	return 0
 }
 
 func (x *ListCategoryItemsReq) GetSort() string {
@@ -2087,8 +2087,8 @@ type ListPromotionItemsReq struct {
 	unknownFields protoimpl.UnknownFields
 
 	Slug string `protobuf:"bytes,1,opt,name=slug,proto3" json:"slug,omitempty"`
-	Page string `protobuf:"bytes,2,opt,name=page,proto3" json:"page,omitempty"`
-	Size string `protobuf:"bytes,3,opt,name=size,proto3" json:"size,omitempty"`
+	Page uint32 `protobuf:"varint,2,opt,name=page,proto3" json:"page,omitempty"`
+	Size uint32 `protobuf:"varint,3,opt,name=size,proto3" json:"size,omitempty"`
 }
 
 func (x *ListPromotionItemsReq) Reset() {
@@ -2130,18 +2130,18 @@ func (x *ListPromotionItemsReq) GetSlug() string {
 	return ""
 }
 
-func (x *ListPromotionItemsReq) GetPage() string {
+func (x *ListPromotionItemsReq) GetPage() uint32 {
 	if x != nil {
 		return x.Page
 	}
-	return ""
+	return 0
 }
 
-func (x *ListPromotionItemsReq) GetSize() string {
+func (x *ListPromotionItemsReq) GetSize() uint32 {
 	if x != nil {
 		return x.Size
 	}
-	return ""
+	return 0
 }
 
 var File_api_pb_products_proto protoreflect.FileDescriptor
@@ -2307,8 +2307,8 @@ var file_api_pb_products_proto_rawDesc = []byte{
 	0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x75, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x41,
 	0x74, 0x22, 0x77, 0x0a, 0x14, 0x6c, 0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72,
 	0x79, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x52, 0x65, 0x71, 0x12, 0x12, 0x0a, 0x04, 0x70, 0x61, 0x67,
-	0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x70, 0x61, 0x67, 0x65, 0x12, 0x12, 0x0a,
-	0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x73, 0x69, 0x7a,
+	0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x70, 0x61, 0x67, 0x65, 0x12, 0x12, 0x0a,
+	0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x73, 0x69, 0x7a,
 	0x65, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x6f, 0x72, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52,
 	0x04, 0x73, 0x6f, 0x72, 0x74, 0x12, 0x23, 0x0a, 0x0d, 0x63, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72,
 	0x79, 0x5f, 0x73, 0x6c, 0x75, 0x67, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x63, 0x61,
@@ -2448,9 +2448,9 @@ var file_api_pb_products_proto_rawDesc = []byte{
 	0x65, 0x22, 0x53, 0x0a, 0x15, 0x4c, 0x69, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69,
 	0x6f, 0x6e, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x52, 0x65, 0x71, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x6c,
 	0x75, 0x67, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x73, 0x6c, 0x75, 0x67, 0x12, 0x12,
-	0x0a, 0x04, 0x70, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x70, 0x61,
-	0x67, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x32, 0xc7, 0x04, 0x0a, 0x04, 0x49, 0x74, 0x65, 0x6d, 0x12,
+	0x0a, 0x04, 0x70, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x70, 0x61,
+	0x67, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d,
+	0x52, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x32, 0xc5, 0x04, 0x0a, 0x04, 0x49, 0x74, 0x65, 0x6d, 0x12,
 	0x35, 0x0a, 0x0a, 0x49, 0x74, 0x65, 0x6d, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x12, 0x0f, 0x2e,
 	0x75, 0x73, 0x65, 0x72, 0x2e, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x52, 0x65, 0x71, 0x1a, 0x16,
 	0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x49,
@@ -2469,95 +2469,95 @@ var file_api_pb_products_proto_rawDesc = []byte{
 	0x65, 0x72, 0x2e, 0x49, 0x74, 0x65, 0x6d, 0x4d, 0x73, 0x67, 0x12, 0x2e, 0x0a, 0x0a, 0x55, 0x70,
 	0x64, 0x61, 0x74, 0x65, 0x49, 0x74, 0x65, 0x6d, 0x12, 0x11, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
 	0x49, 0x74, 0x65, 0x6d, 0x57, 0x69, 0x74, 0x68, 0x55, 0x69, 0x64, 0x1a, 0x0d, 0x2e, 0x75, 0x73,
-	0x65, 0x72, 0x2e, 0x49, 0x74, 0x65, 0x6d, 0x4d, 0x73, 0x67, 0x12, 0x2a, 0x0a, 0x0a, 0x44, 0x65,
+	0x65, 0x72, 0x2e, 0x49, 0x74, 0x65, 0x6d, 0x4d, 0x73, 0x67, 0x12, 0x28, 0x0a, 0x0a, 0x44, 0x65,
 	0x6c, 0x65, 0x74, 0x65, 0x49, 0x74, 0x65, 0x6d, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
-	0x75, 0x75, 0x69, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x49,
-	0x74, 0x65, 0x6d, 0x4d, 0x73, 0x67, 0x12, 0x39, 0x0a, 0x10, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65,
-	0x6c, 0x61, 0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65,
-	0x72, 0x2e, 0x75, 0x75, 0x69, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72,
-	0x2e, 0x52, 0x65, 0x6c, 0x61, 0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x4c, 0x69, 0x73,
-	0x74, 0x12, 0x47, 0x0a, 0x11, 0x6c, 0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72,
-	0x79, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x1a, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x6c, 0x69,
-	0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x52,
-	0x65, 0x71, 0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61,
-	0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x52, 0x65, 0x73, 0x12, 0x31, 0x0a, 0x08, 0x48, 0x69,
-	0x74, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c, 0x69,
-	0x73, 0x74, 0x52, 0x65, 0x71, 0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67,
-	0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x52, 0x65, 0x73, 0x12, 0x31, 0x0a,
-	0x08, 0x52, 0x65, 0x63, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72,
-	0x2e, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
-	0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x52, 0x65, 0x73,
-	0x32, 0xd9, 0x03, 0x0a, 0x08, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x3b, 0x0a,
-	0x0e, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x69, 0x65, 0x73, 0x12,
-	0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x1a, 0x1a,
-	0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x43,
-	0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x52, 0x65, 0x73, 0x12, 0x36, 0x0a, 0x0e, 0x43, 0x72,
-	0x65, 0x61, 0x74, 0x65, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x11, 0x2e, 0x75,
-	0x73, 0x65, 0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4d, 0x73, 0x67, 0x1a,
-	0x11, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4d,
-	0x73, 0x67, 0x12, 0x3d, 0x0a, 0x0e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x53, 0x65,
-	0x61, 0x72, 0x63, 0x68, 0x12, 0x0f, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x53, 0x65, 0x61, 0x72,
-	0x63, 0x68, 0x52, 0x65, 0x71, 0x1a, 0x1a, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67,
-	0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x52, 0x65,
-	0x73, 0x12, 0x42, 0x0a, 0x15, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x46, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x73, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x12, 0x0f, 0x2e, 0x75, 0x73, 0x65,
-	0x72, 0x2e, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x52, 0x65, 0x71, 0x1a, 0x18, 0x2e, 0x75, 0x73,
-	0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x46, 0x69, 0x6c, 0x74,
-	0x65, 0x72, 0x52, 0x65, 0x73, 0x12, 0x2f, 0x0a, 0x0b, 0x47, 0x65, 0x74, 0x43, 0x61, 0x74, 0x65,
-	0x67, 0x6f, 0x72, 0x79, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x73, 0x6c, 0x75, 0x67,
-	0x4d, 0x73, 0x67, 0x1a, 0x11, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67,
-	0x6f, 0x72, 0x79, 0x4d, 0x73, 0x67, 0x12, 0x3b, 0x0a, 0x0e, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65,
-	0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
-	0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x57, 0x69, 0x74, 0x68, 0x53, 0x6c, 0x75, 0x67,
-	0x1a, 0x11, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79,
-	0x4d, 0x73, 0x67, 0x12, 0x2c, 0x0a, 0x0e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x43, 0x61, 0x74,
-	0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x73, 0x6c, 0x75,
+	0x75, 0x75, 0x69, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x0b, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x45,
+	0x6d, 0x70, 0x74, 0x79, 0x12, 0x39, 0x0a, 0x10, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x6c, 0x61,
+	0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
+	0x75, 0x75, 0x69, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x52,
+	0x65, 0x6c, 0x61, 0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x4c, 0x69, 0x73, 0x74, 0x12,
+	0x47, 0x0a, 0x11, 0x6c, 0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x49,
+	0x74, 0x65, 0x6d, 0x73, 0x12, 0x1a, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x6c, 0x69, 0x73, 0x74,
+	0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x52, 0x65, 0x71,
+	0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65,
+	0x64, 0x49, 0x74, 0x65, 0x6d, 0x52, 0x65, 0x73, 0x12, 0x31, 0x0a, 0x08, 0x48, 0x69, 0x74, 0x49,
+	0x74, 0x65, 0x6d, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c, 0x69, 0x73, 0x74,
+	0x52, 0x65, 0x71, 0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e,
+	0x61, 0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x52, 0x65, 0x73, 0x12, 0x31, 0x0a, 0x08, 0x52,
+	0x65, 0x63, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c,
+	0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x1a, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61,
+	0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x49, 0x74, 0x65, 0x6d, 0x52, 0x65, 0x73, 0x32, 0xd9,
+	0x03, 0x0a, 0x08, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x3b, 0x0a, 0x0e, 0x4c,
+	0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x69, 0x65, 0x73, 0x12, 0x0d, 0x2e,
+	0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x1a, 0x1a, 0x2e, 0x75,
+	0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x43, 0x61, 0x74,
+	0x65, 0x67, 0x6f, 0x72, 0x79, 0x52, 0x65, 0x73, 0x12, 0x36, 0x0a, 0x0e, 0x43, 0x72, 0x65, 0x61,
+	0x74, 0x65, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x11, 0x2e, 0x75, 0x73, 0x65,
+	0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4d, 0x73, 0x67, 0x1a, 0x11, 0x2e,
+	0x75, 0x73, 0x65, 0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4d, 0x73, 0x67,
+	0x12, 0x3d, 0x0a, 0x0e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x53, 0x65, 0x61, 0x72,
+	0x63, 0x68, 0x12, 0x0f, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68,
+	0x52, 0x65, 0x71, 0x1a, 0x1a, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e,
+	0x61, 0x74, 0x65, 0x64, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x52, 0x65, 0x73, 0x12,
+	0x42, 0x0a, 0x15, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x46, 0x69, 0x6c, 0x74, 0x65,
+	0x72, 0x73, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x12, 0x0f, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
+	0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x52, 0x65, 0x71, 0x1a, 0x18, 0x2e, 0x75, 0x73, 0x65, 0x72,
+	0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72,
+	0x52, 0x65, 0x73, 0x12, 0x2f, 0x0a, 0x0b, 0x47, 0x65, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f,
+	0x72, 0x79, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x73, 0x6c, 0x75, 0x67, 0x4d, 0x73,
+	0x67, 0x1a, 0x11, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72,
+	0x79, 0x4d, 0x73, 0x67, 0x12, 0x3b, 0x0a, 0x0e, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x43, 0x61,
+	0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x16, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x43, 0x61,
+	0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x57, 0x69, 0x74, 0x68, 0x53, 0x6c, 0x75, 0x67, 0x1a, 0x11,
+	0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4d, 0x73,
+	0x67, 0x12, 0x2c, 0x0a, 0x0e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x43, 0x61, 0x74, 0x65, 0x67,
+	0x6f, 0x72, 0x79, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x73, 0x6c, 0x75, 0x67, 0x4d,
+	0x73, 0x67, 0x1a, 0x0b, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12,
+	0x39, 0x0a, 0x13, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x46,
+	0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x73, 0x6c,
+	0x75, 0x67, 0x4d, 0x73, 0x67, 0x1a, 0x13, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x46, 0x69, 0x6c,
+	0x74, 0x65, 0x72, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x73, 0x32, 0xb6, 0x01, 0x0a, 0x08, 0x46,
+	0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x12, 0x35, 0x0a, 0x0d, 0x4c, 0x69, 0x73, 0x74, 0x46,
+	0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
+	0x75, 0x75, 0x69, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x15, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x46,
+	0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x4d, 0x73, 0x67, 0x12, 0x39,
+	0x0a, 0x0e, 0x41, 0x64, 0x64, 0x54, 0x6f, 0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x73,
+	0x12, 0x14, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x41, 0x6e, 0x64, 0x49,
+	0x74, 0x65, 0x6d, 0x49, 0x64, 0x73, 0x1a, 0x11, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x46, 0x61,
+	0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x4d, 0x73, 0x67, 0x12, 0x38, 0x0a, 0x13, 0x52, 0x65, 0x6d,
+	0x6f, 0x76, 0x65, 0x46, 0x72, 0x6f, 0x6d, 0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x73,
+	0x12, 0x14, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x41, 0x6e, 0x64, 0x49,
+	0x74, 0x65, 0x6d, 0x49, 0x64, 0x73, 0x1a, 0x0b, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x45, 0x6d,
+	0x70, 0x74, 0x79, 0x32, 0x9c, 0x03, 0x0a, 0x09, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f,
+	0x6e, 0x12, 0x38, 0x0a, 0x0e, 0x4c, 0x69, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69,
+	0x6f, 0x6e, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x52,
+	0x65, 0x71, 0x1a, 0x17, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61,
+	0x74, 0x65, 0x64, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x52, 0x65, 0x73, 0x12, 0x3b, 0x0a, 0x0f, 0x50,
+	0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x12, 0x0f,
+	0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x52, 0x65, 0x71, 0x1a,
+	0x17, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64,
+	0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x52, 0x65, 0x73, 0x12, 0x31, 0x0a, 0x0f, 0x43, 0x72, 0x65, 0x61,
+	0x74, 0x65, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0e, 0x2e, 0x75, 0x73,
+	0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x4d, 0x73, 0x67, 0x1a, 0x0e, 0x2e, 0x75, 0x73,
+	0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x4d, 0x73, 0x67, 0x12, 0x2d, 0x0a, 0x0c, 0x47,
+	0x65, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0d, 0x2e, 0x75, 0x73,
+	0x65, 0x72, 0x2e, 0x73, 0x6c, 0x75, 0x67, 0x4d, 0x73, 0x67, 0x1a, 0x0e, 0x2e, 0x75, 0x73, 0x65,
+	0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x4d, 0x73, 0x67, 0x12, 0x36, 0x0a, 0x0f, 0x55, 0x70,
+	0x64, 0x61, 0x74, 0x65, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x13, 0x2e,
+	0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x57, 0x69, 0x74, 0x68, 0x53, 0x6c,
+	0x75, 0x67, 0x1a, 0x0e, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x4d,
+	0x73, 0x67, 0x12, 0x2d, 0x0a, 0x0f, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x50, 0x72, 0x6f, 0x6d,
+	0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x73, 0x6c, 0x75,
 	0x67, 0x4d, 0x73, 0x67, 0x1a, 0x0b, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x45, 0x6d, 0x70, 0x74,
-	0x79, 0x12, 0x39, 0x0a, 0x13, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72,
-	0x79, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
-	0x73, 0x6c, 0x75, 0x67, 0x4d, 0x73, 0x67, 0x1a, 0x13, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x46,
-	0x69, 0x6c, 0x74, 0x65, 0x72, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x73, 0x32, 0xb6, 0x01, 0x0a,
-	0x08, 0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x12, 0x35, 0x0a, 0x0d, 0x4c, 0x69, 0x73,
-	0x74, 0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65,
-	0x72, 0x2e, 0x75, 0x75, 0x69, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x15, 0x2e, 0x75, 0x73, 0x65, 0x72,
-	0x2e, 0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x4d, 0x73, 0x67,
-	0x12, 0x39, 0x0a, 0x0e, 0x41, 0x64, 0x64, 0x54, 0x6f, 0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74,
-	0x65, 0x73, 0x12, 0x14, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x41, 0x6e,
-	0x64, 0x49, 0x74, 0x65, 0x6d, 0x49, 0x64, 0x73, 0x1a, 0x11, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
-	0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74, 0x65, 0x4d, 0x73, 0x67, 0x12, 0x38, 0x0a, 0x13, 0x52,
-	0x65, 0x6d, 0x6f, 0x76, 0x65, 0x46, 0x72, 0x6f, 0x6d, 0x46, 0x61, 0x76, 0x6f, 0x72, 0x69, 0x74,
-	0x65, 0x73, 0x12, 0x14, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x41, 0x6e,
-	0x64, 0x49, 0x74, 0x65, 0x6d, 0x49, 0x64, 0x73, 0x1a, 0x0b, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e,
-	0x45, 0x6d, 0x70, 0x74, 0x79, 0x32, 0x9c, 0x03, 0x0a, 0x09, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74,
-	0x69, 0x6f, 0x6e, 0x12, 0x38, 0x0a, 0x0e, 0x4c, 0x69, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f,
-	0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c, 0x69, 0x73,
-	0x74, 0x52, 0x65, 0x71, 0x1a, 0x17, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69,
-	0x6e, 0x61, 0x74, 0x65, 0x64, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x52, 0x65, 0x73, 0x12, 0x3b, 0x0a,
-	0x0f, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68,
-	0x12, 0x0f, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x53, 0x65, 0x61, 0x72, 0x63, 0x68, 0x52, 0x65,
-	0x71, 0x1a, 0x17, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74,
-	0x65, 0x64, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x52, 0x65, 0x73, 0x12, 0x31, 0x0a, 0x0f, 0x43, 0x72,
-	0x65, 0x61, 0x74, 0x65, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0e, 0x2e,
-	0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x4d, 0x73, 0x67, 0x1a, 0x0e, 0x2e,
-	0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x4d, 0x73, 0x67, 0x12, 0x2d, 0x0a,
-	0x0c, 0x47, 0x65, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0d, 0x2e,
-	0x75, 0x73, 0x65, 0x72, 0x2e, 0x73, 0x6c, 0x75, 0x67, 0x4d, 0x73, 0x67, 0x1a, 0x0e, 0x2e, 0x75,
-	0x73, 0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x4d, 0x73, 0x67, 0x12, 0x36, 0x0a, 0x0f,
-	0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12,
-	0x13, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x57, 0x69, 0x74, 0x68,
-	0x53, 0x6c, 0x75, 0x67, 0x1a, 0x0e, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x72, 0x6f, 0x6d,
-	0x6f, 0x4d, 0x73, 0x67, 0x12, 0x2d, 0x0a, 0x0f, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x50, 0x72,
-	0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x0d, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x73,
-	0x6c, 0x75, 0x67, 0x4d, 0x73, 0x67, 0x1a, 0x0b, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x45, 0x6d,
-	0x70, 0x74, 0x79, 0x12, 0x4f, 0x0a, 0x12, 0x4c, 0x69, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f,
-	0x74, 0x69, 0x6f, 0x6e, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x1b, 0x2e, 0x75, 0x73, 0x65, 0x72,
-	0x2e, 0x4c, 0x69, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x74,
-	0x65, 0x6d, 0x73, 0x52, 0x65, 0x71, 0x1a, 0x1c, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61,
-	0x67, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x64, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x49, 0x74, 0x65, 0x6d,
-	0x73, 0x52, 0x65, 0x73, 0x42, 0x27, 0x5a, 0x25, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63,
-	0x6f, 0x6d, 0x2f, 0x4a, 0x4d, 0x55, 0x52, 0x76, 0x2f, 0x70, 0x61, 0x72, 0x2d, 0x70, 0x72, 0x6f,
-	0x2f, 0x61, 0x70, 0x69, 0x2f, 0x70, 0x62, 0x2f, 0x75, 0x73, 0x65, 0x72, 0x73, 0x62, 0x06, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x79, 0x12, 0x4f, 0x0a, 0x12, 0x4c, 0x69, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69,
+	0x6f, 0x6e, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x12, 0x1b, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x4c,
+	0x69, 0x73, 0x74, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x74, 0x65, 0x6d,
+	0x73, 0x52, 0x65, 0x71, 0x1a, 0x1c, 0x2e, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x50, 0x61, 0x67, 0x69,
+	0x6e, 0x61, 0x74, 0x65, 0x64, 0x50, 0x72, 0x6f, 0x6d, 0x6f, 0x49, 0x74, 0x65, 0x6d, 0x73, 0x52,
+	0x65, 0x73, 0x42, 0x27, 0x5a, 0x25, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
+	0x2f, 0x4a, 0x4d, 0x55, 0x52, 0x76, 0x2f, 0x70, 0x61, 0x72, 0x2d, 0x70, 0x72, 0x6f, 0x2f, 0x61,
+	0x70, 0x69, 0x2f, 0x70, 0x62, 0x2f, 0x75, 0x73, 0x65, 0x72, 0x73, 0x62, 0x06, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x33,
 }
 
 var (
@@ -2573,7 +2573,7 @@ func file_api_pb_products_proto_rawDescGZIP() []byte {
 }
 
 var file_api_pb_products_proto_msgTypes = make([]protoimpl.MessageInfo, 28)
-var file_api_pb_products_proto_goTypes = []interface{}{
+var file_api_pb_products_proto_goTypes = []any{
 	(*Empty)(nil),                  // 0: user.Empty
 	(*UuidMsg)(nil),                // 1: user.uuidMsg
 	(*SlugMsg)(nil),                // 2: user.slugMsg
@@ -2682,7 +2682,7 @@ var file_api_pb_products_proto_depIdxs = []int32{
 	8,  // 74: user.Item.CreateItem:output_type -> user.ItemMsg
 	8,  // 75: user.Item.GetItem:output_type -> user.ItemMsg
 	8,  // 76: user.Item.UpdateItem:output_type -> user.ItemMsg
-	8,  // 77: user.Item.DeleteItem:output_type -> user.ItemMsg
+	0,  // 77: user.Item.DeleteItem:output_type -> user.Empty
 	13, // 78: user.Item.ListRelatedItems:output_type -> user.RelatedItemsList
 	15, // 79: user.Item.listCategoryItems:output_type -> user.PaginatedItemRes
 	15, // 80: user.Item.HitItems:output_type -> user.PaginatedItemRes
@@ -2718,7 +2718,7 @@ func file_api_pb_products_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_api_pb_products_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Empty); i {
 			case 0:
 				return &v.state
@@ -2730,7 +2730,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*UuidMsg); i {
 			case 0:
 				return &v.state
@@ -2742,7 +2742,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*SlugMsg); i {
 			case 0:
 				return &v.state
@@ -2754,7 +2754,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ListReq); i {
 			case 0:
 				return &v.state
@@ -2766,7 +2766,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*SearchReq); i {
 			case 0:
 				return &v.state
@@ -2778,7 +2778,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*CategoryMsg); i {
 			case 0:
 				return &v.state
@@ -2790,7 +2790,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*CategoryWithSlug); i {
 			case 0:
 				return &v.state
@@ -2802,7 +2802,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*Filter); i {
 			case 0:
 				return &v.state
@@ -2814,7 +2814,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*ItemMsg); i {
 			case 0:
 				return &v.state
@@ -2826,7 +2826,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*ItemMedia); i {
 			case 0:
 				return &v.state
@@ -2838,7 +2838,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*ItemAttribute); i {
 			case 0:
 				return &v.state
@@ -2850,7 +2850,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*RelatedProduct); i {
 			case 0:
 				return &v.state
@@ -2862,7 +2862,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*ListCategoryItemsReq); i {
 			case 0:
 				return &v.state
@@ -2874,7 +2874,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*RelatedItemsList); i {
 			case 0:
 				return &v.state
@@ -2886,7 +2886,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*ItemWithUid); i {
 			case 0:
 				return &v.state
@@ -2898,7 +2898,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*PaginatedItemRes); i {
 			case 0:
 				return &v.state
@@ -2910,7 +2910,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[16].Exporter = func(v any, i int) any {
 			switch v := v.(*PaginatedCategoryRes); i {
 			case 0:
 				return &v.state
@@ -2922,7 +2922,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[17].Exporter = func(v any, i int) any {
 			switch v := v.(*FilterListRes); i {
 			case 0:
 				return &v.state
@@ -2934,7 +2934,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[18].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[18].Exporter = func(v any, i int) any {
 			switch v := v.(*PaginatedFilterRes); i {
 			case 0:
 				return &v.state
@@ -2946,7 +2946,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[19].Exporter = func(v any, i int) any {
 			switch v := v.(*FavoriteMsg); i {
 			case 0:
 				return &v.state
@@ -2958,7 +2958,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[20].Exporter = func(v any, i int) any {
 			switch v := v.(*FavoriteListMsg); i {
 			case 0:
 				return &v.state
@@ -2970,7 +2970,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[21].Exporter = func(v any, i int) any {
 			switch v := v.(*UserAndItemIds); i {
 			case 0:
 				return &v.state
@@ -2982,7 +2982,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[22].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[22].Exporter = func(v any, i int) any {
 			switch v := v.(*PromoMsg); i {
 			case 0:
 				return &v.state
@@ -2994,7 +2994,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[23].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[23].Exporter = func(v any, i int) any {
 			switch v := v.(*PromoWithSlug); i {
 			case 0:
 				return &v.state
@@ -3006,7 +3006,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[24].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[24].Exporter = func(v any, i int) any {
 			switch v := v.(*PromoItem); i {
 			case 0:
 				return &v.state
@@ -3018,7 +3018,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[25].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[25].Exporter = func(v any, i int) any {
 			switch v := v.(*PaginatedPromoRes); i {
 			case 0:
 				return &v.state
@@ -3030,7 +3030,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[26].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[26].Exporter = func(v any, i int) any {
 			switch v := v.(*PaginatedPromoItemsRes); i {
 			case 0:
 				return &v.state
@@ -3042,7 +3042,7 @@ func file_api_pb_products_proto_init() {
 				return nil
 			}
 		}
-		file_api_pb_products_proto_msgTypes[27].Exporter = func(v interface{}, i int) interface{} {
+		file_api_pb_products_proto_msgTypes[27].Exporter = func(v any, i int) any {
 			switch v := v.(*ListPromotionItemsReq); i {
 			case 0:
 				return &v.state
