@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             v4.25.1
 // source: api/pb/products.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	Item_ItemSearch_FullMethodName        = "/user.Item/ItemSearch"
@@ -42,7 +42,7 @@ type ItemClient interface {
 	CreateItem(ctx context.Context, in *ItemMsg, opts ...grpc.CallOption) (*ItemMsg, error)
 	GetItem(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*ItemMsg, error)
 	UpdateItem(ctx context.Context, in *ItemWithUid, opts ...grpc.CallOption) (*ItemMsg, error)
-	DeleteItem(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*ItemMsg, error)
+	DeleteItem(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*Empty, error)
 	ListRelatedItems(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*RelatedItemsList, error)
 	ListCategoryItems(ctx context.Context, in *ListCategoryItemsReq, opts ...grpc.CallOption) (*PaginatedItemRes, error)
 	HitItems(ctx context.Context, in *ListReq, opts ...grpc.CallOption) (*PaginatedItemRes, error)
@@ -58,8 +58,9 @@ func NewItemClient(cc grpc.ClientConnInterface) ItemClient {
 }
 
 func (c *itemClient) ItemSearch(ctx context.Context, in *SearchReq, opts ...grpc.CallOption) (*PaginatedItemRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedItemRes)
-	err := c.cc.Invoke(ctx, Item_ItemSearch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_ItemSearch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -67,8 +68,9 @@ func (c *itemClient) ItemSearch(ctx context.Context, in *SearchReq, opts ...grpc
 }
 
 func (c *itemClient) ItemAttrSearch(ctx context.Context, in *SearchReq, opts ...grpc.CallOption) (*PaginatedItemRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedItemRes)
-	err := c.cc.Invoke(ctx, Item_ItemAttrSearch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_ItemAttrSearch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -76,8 +78,9 @@ func (c *itemClient) ItemAttrSearch(ctx context.Context, in *SearchReq, opts ...
 }
 
 func (c *itemClient) ListItems(ctx context.Context, in *ListReq, opts ...grpc.CallOption) (*PaginatedItemRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedItemRes)
-	err := c.cc.Invoke(ctx, Item_ListItems_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_ListItems_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -85,8 +88,9 @@ func (c *itemClient) ListItems(ctx context.Context, in *ListReq, opts ...grpc.Ca
 }
 
 func (c *itemClient) CreateItem(ctx context.Context, in *ItemMsg, opts ...grpc.CallOption) (*ItemMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ItemMsg)
-	err := c.cc.Invoke(ctx, Item_CreateItem_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_CreateItem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -94,8 +98,9 @@ func (c *itemClient) CreateItem(ctx context.Context, in *ItemMsg, opts ...grpc.C
 }
 
 func (c *itemClient) GetItem(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*ItemMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ItemMsg)
-	err := c.cc.Invoke(ctx, Item_GetItem_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_GetItem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -103,17 +108,19 @@ func (c *itemClient) GetItem(ctx context.Context, in *UuidMsg, opts ...grpc.Call
 }
 
 func (c *itemClient) UpdateItem(ctx context.Context, in *ItemWithUid, opts ...grpc.CallOption) (*ItemMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ItemMsg)
-	err := c.cc.Invoke(ctx, Item_UpdateItem_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_UpdateItem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *itemClient) DeleteItem(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*ItemMsg, error) {
-	out := new(ItemMsg)
-	err := c.cc.Invoke(ctx, Item_DeleteItem_FullMethodName, in, out, opts...)
+func (c *itemClient) DeleteItem(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(Empty)
+	err := c.cc.Invoke(ctx, Item_DeleteItem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -121,8 +128,9 @@ func (c *itemClient) DeleteItem(ctx context.Context, in *UuidMsg, opts ...grpc.C
 }
 
 func (c *itemClient) ListRelatedItems(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*RelatedItemsList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(RelatedItemsList)
-	err := c.cc.Invoke(ctx, Item_ListRelatedItems_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_ListRelatedItems_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -130,8 +138,9 @@ func (c *itemClient) ListRelatedItems(ctx context.Context, in *UuidMsg, opts ...
 }
 
 func (c *itemClient) ListCategoryItems(ctx context.Context, in *ListCategoryItemsReq, opts ...grpc.CallOption) (*PaginatedItemRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedItemRes)
-	err := c.cc.Invoke(ctx, Item_ListCategoryItems_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_ListCategoryItems_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -139,8 +148,9 @@ func (c *itemClient) ListCategoryItems(ctx context.Context, in *ListCategoryItem
 }
 
 func (c *itemClient) HitItems(ctx context.Context, in *ListReq, opts ...grpc.CallOption) (*PaginatedItemRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedItemRes)
-	err := c.cc.Invoke(ctx, Item_HitItems_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_HitItems_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -148,8 +158,9 @@ func (c *itemClient) HitItems(ctx context.Context, in *ListReq, opts ...grpc.Cal
 }
 
 func (c *itemClient) RecItems(ctx context.Context, in *ListReq, opts ...grpc.CallOption) (*PaginatedItemRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedItemRes)
-	err := c.cc.Invoke(ctx, Item_RecItems_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Item_RecItems_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -158,7 +169,7 @@ func (c *itemClient) RecItems(ctx context.Context, in *ListReq, opts ...grpc.Cal
 
 // ItemServer is the server API for Item service.
 // All implementations must embed UnimplementedItemServer
-// for forward compatibility
+// for forward compatibility.
 type ItemServer interface {
 	ItemSearch(context.Context, *SearchReq) (*PaginatedItemRes, error)
 	ItemAttrSearch(context.Context, *SearchReq) (*PaginatedItemRes, error)
@@ -166,7 +177,7 @@ type ItemServer interface {
 	CreateItem(context.Context, *ItemMsg) (*ItemMsg, error)
 	GetItem(context.Context, *UuidMsg) (*ItemMsg, error)
 	UpdateItem(context.Context, *ItemWithUid) (*ItemMsg, error)
-	DeleteItem(context.Context, *UuidMsg) (*ItemMsg, error)
+	DeleteItem(context.Context, *UuidMsg) (*Empty, error)
 	ListRelatedItems(context.Context, *UuidMsg) (*RelatedItemsList, error)
 	ListCategoryItems(context.Context, *ListCategoryItemsReq) (*PaginatedItemRes, error)
 	HitItems(context.Context, *ListReq) (*PaginatedItemRes, error)
@@ -174,9 +185,12 @@ type ItemServer interface {
 	mustEmbedUnimplementedItemServer()
 }
 
-// UnimplementedItemServer must be embedded to have forward compatible implementations.
-type UnimplementedItemServer struct {
-}
+// UnimplementedItemServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedItemServer struct{}
 
 func (UnimplementedItemServer) ItemSearch(context.Context, *SearchReq) (*PaginatedItemRes, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ItemSearch not implemented")
@@ -196,7 +210,7 @@ func (UnimplementedItemServer) GetItem(context.Context, *UuidMsg) (*ItemMsg, err
 func (UnimplementedItemServer) UpdateItem(context.Context, *ItemWithUid) (*ItemMsg, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateItem not implemented")
 }
-func (UnimplementedItemServer) DeleteItem(context.Context, *UuidMsg) (*ItemMsg, error) {
+func (UnimplementedItemServer) DeleteItem(context.Context, *UuidMsg) (*Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DeleteItem not implemented")
 }
 func (UnimplementedItemServer) ListRelatedItems(context.Context, *UuidMsg) (*RelatedItemsList, error) {
@@ -212,6 +226,7 @@ func (UnimplementedItemServer) RecItems(context.Context, *ListReq) (*PaginatedIt
 	return nil, status.Errorf(codes.Unimplemented, "method RecItems not implemented")
 }
 func (UnimplementedItemServer) mustEmbedUnimplementedItemServer() {}
+func (UnimplementedItemServer) testEmbeddedByValue()              {}
 
 // UnsafeItemServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ItemServer will
@@ -221,6 +236,13 @@ type UnsafeItemServer interface {
 }
 
 func RegisterItemServer(s grpc.ServiceRegistrar, srv ItemServer) {
+	// If the following call pancis, it indicates UnimplementedItemServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Item_ServiceDesc, srv)
 }
 
@@ -512,8 +534,9 @@ func NewCategoryClient(cc grpc.ClientConnInterface) CategoryClient {
 }
 
 func (c *categoryClient) ListCategories(ctx context.Context, in *ListReq, opts ...grpc.CallOption) (*PaginatedCategoryRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedCategoryRes)
-	err := c.cc.Invoke(ctx, Category_ListCategories_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_ListCategories_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -521,8 +544,9 @@ func (c *categoryClient) ListCategories(ctx context.Context, in *ListReq, opts .
 }
 
 func (c *categoryClient) CreateCategory(ctx context.Context, in *CategoryMsg, opts ...grpc.CallOption) (*CategoryMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CategoryMsg)
-	err := c.cc.Invoke(ctx, Category_CreateCategory_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_CreateCategory_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -530,8 +554,9 @@ func (c *categoryClient) CreateCategory(ctx context.Context, in *CategoryMsg, op
 }
 
 func (c *categoryClient) CategorySearch(ctx context.Context, in *SearchReq, opts ...grpc.CallOption) (*PaginatedCategoryRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedCategoryRes)
-	err := c.cc.Invoke(ctx, Category_CategorySearch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_CategorySearch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -539,8 +564,9 @@ func (c *categoryClient) CategorySearch(ctx context.Context, in *SearchReq, opts
 }
 
 func (c *categoryClient) CategoryFiltersSearch(ctx context.Context, in *SearchReq, opts ...grpc.CallOption) (*PaginatedFilterRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedFilterRes)
-	err := c.cc.Invoke(ctx, Category_CategoryFiltersSearch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_CategoryFiltersSearch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -548,8 +574,9 @@ func (c *categoryClient) CategoryFiltersSearch(ctx context.Context, in *SearchRe
 }
 
 func (c *categoryClient) GetCategory(ctx context.Context, in *SlugMsg, opts ...grpc.CallOption) (*CategoryMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CategoryMsg)
-	err := c.cc.Invoke(ctx, Category_GetCategory_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_GetCategory_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -557,8 +584,9 @@ func (c *categoryClient) GetCategory(ctx context.Context, in *SlugMsg, opts ...g
 }
 
 func (c *categoryClient) UpdateCategory(ctx context.Context, in *CategoryWithSlug, opts ...grpc.CallOption) (*CategoryMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CategoryMsg)
-	err := c.cc.Invoke(ctx, Category_UpdateCategory_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_UpdateCategory_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -566,8 +594,9 @@ func (c *categoryClient) UpdateCategory(ctx context.Context, in *CategoryWithSlu
 }
 
 func (c *categoryClient) DeleteCategory(ctx context.Context, in *SlugMsg, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, Category_DeleteCategory_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_DeleteCategory_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -575,8 +604,9 @@ func (c *categoryClient) DeleteCategory(ctx context.Context, in *SlugMsg, opts .
 }
 
 func (c *categoryClient) ListCategoryFilters(ctx context.Context, in *SlugMsg, opts ...grpc.CallOption) (*FilterListRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FilterListRes)
-	err := c.cc.Invoke(ctx, Category_ListCategoryFilters_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Category_ListCategoryFilters_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -585,7 +615,7 @@ func (c *categoryClient) ListCategoryFilters(ctx context.Context, in *SlugMsg, o
 
 // CategoryServer is the server API for Category service.
 // All implementations must embed UnimplementedCategoryServer
-// for forward compatibility
+// for forward compatibility.
 type CategoryServer interface {
 	ListCategories(context.Context, *ListReq) (*PaginatedCategoryRes, error)
 	CreateCategory(context.Context, *CategoryMsg) (*CategoryMsg, error)
@@ -598,9 +628,12 @@ type CategoryServer interface {
 	mustEmbedUnimplementedCategoryServer()
 }
 
-// UnimplementedCategoryServer must be embedded to have forward compatible implementations.
-type UnimplementedCategoryServer struct {
-}
+// UnimplementedCategoryServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedCategoryServer struct{}
 
 func (UnimplementedCategoryServer) ListCategories(context.Context, *ListReq) (*PaginatedCategoryRes, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListCategories not implemented")
@@ -627,6 +660,7 @@ func (UnimplementedCategoryServer) ListCategoryFilters(context.Context, *SlugMsg
 	return nil, status.Errorf(codes.Unimplemented, "method ListCategoryFilters not implemented")
 }
 func (UnimplementedCategoryServer) mustEmbedUnimplementedCategoryServer() {}
+func (UnimplementedCategoryServer) testEmbeddedByValue()                  {}
 
 // UnsafeCategoryServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to CategoryServer will
@@ -636,6 +670,13 @@ type UnsafeCategoryServer interface {
 }
 
 func RegisterCategoryServer(s grpc.ServiceRegistrar, srv CategoryServer) {
+	// If the following call pancis, it indicates UnimplementedCategoryServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Category_ServiceDesc, srv)
 }
 
@@ -851,8 +892,9 @@ func NewFavoriteClient(cc grpc.ClientConnInterface) FavoriteClient {
 }
 
 func (c *favoriteClient) ListFavorites(ctx context.Context, in *UuidMsg, opts ...grpc.CallOption) (*FavoriteListMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FavoriteListMsg)
-	err := c.cc.Invoke(ctx, Favorite_ListFavorites_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Favorite_ListFavorites_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -860,8 +902,9 @@ func (c *favoriteClient) ListFavorites(ctx context.Context, in *UuidMsg, opts ..
 }
 
 func (c *favoriteClient) AddToFavorites(ctx context.Context, in *UserAndItemIds, opts ...grpc.CallOption) (*FavoriteMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FavoriteMsg)
-	err := c.cc.Invoke(ctx, Favorite_AddToFavorites_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Favorite_AddToFavorites_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -869,8 +912,9 @@ func (c *favoriteClient) AddToFavorites(ctx context.Context, in *UserAndItemIds,
 }
 
 func (c *favoriteClient) RemoveFromFavorites(ctx context.Context, in *UserAndItemIds, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, Favorite_RemoveFromFavorites_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Favorite_RemoveFromFavorites_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -879,7 +923,7 @@ func (c *favoriteClient) RemoveFromFavorites(ctx context.Context, in *UserAndIte
 
 // FavoriteServer is the server API for Favorite service.
 // All implementations must embed UnimplementedFavoriteServer
-// for forward compatibility
+// for forward compatibility.
 type FavoriteServer interface {
 	ListFavorites(context.Context, *UuidMsg) (*FavoriteListMsg, error)
 	AddToFavorites(context.Context, *UserAndItemIds) (*FavoriteMsg, error)
@@ -887,9 +931,12 @@ type FavoriteServer interface {
 	mustEmbedUnimplementedFavoriteServer()
 }
 
-// UnimplementedFavoriteServer must be embedded to have forward compatible implementations.
-type UnimplementedFavoriteServer struct {
-}
+// UnimplementedFavoriteServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedFavoriteServer struct{}
 
 func (UnimplementedFavoriteServer) ListFavorites(context.Context, *UuidMsg) (*FavoriteListMsg, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListFavorites not implemented")
@@ -901,6 +948,7 @@ func (UnimplementedFavoriteServer) RemoveFromFavorites(context.Context, *UserAnd
 	return nil, status.Errorf(codes.Unimplemented, "method RemoveFromFavorites not implemented")
 }
 func (UnimplementedFavoriteServer) mustEmbedUnimplementedFavoriteServer() {}
+func (UnimplementedFavoriteServer) testEmbeddedByValue()                  {}
 
 // UnsafeFavoriteServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to FavoriteServer will
@@ -910,6 +958,13 @@ type UnsafeFavoriteServer interface {
 }
 
 func RegisterFavoriteServer(s grpc.ServiceRegistrar, srv FavoriteServer) {
+	// If the following call pancis, it indicates UnimplementedFavoriteServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Favorite_ServiceDesc, srv)
 }
 
@@ -1023,8 +1078,9 @@ func NewPromotionClient(cc grpc.ClientConnInterface) PromotionClient {
 }
 
 func (c *promotionClient) ListPromotions(ctx context.Context, in *ListReq, opts ...grpc.CallOption) (*PaginatedPromoRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedPromoRes)
-	err := c.cc.Invoke(ctx, Promotion_ListPromotions_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Promotion_ListPromotions_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1032,8 +1088,9 @@ func (c *promotionClient) ListPromotions(ctx context.Context, in *ListReq, opts 
 }
 
 func (c *promotionClient) PromotionSearch(ctx context.Context, in *SearchReq, opts ...grpc.CallOption) (*PaginatedPromoRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedPromoRes)
-	err := c.cc.Invoke(ctx, Promotion_PromotionSearch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Promotion_PromotionSearch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1041,8 +1098,9 @@ func (c *promotionClient) PromotionSearch(ctx context.Context, in *SearchReq, op
 }
 
 func (c *promotionClient) CreatePromotion(ctx context.Context, in *PromoMsg, opts ...grpc.CallOption) (*PromoMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PromoMsg)
-	err := c.cc.Invoke(ctx, Promotion_CreatePromotion_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Promotion_CreatePromotion_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1050,8 +1108,9 @@ func (c *promotionClient) CreatePromotion(ctx context.Context, in *PromoMsg, opt
 }
 
 func (c *promotionClient) GetPromotion(ctx context.Context, in *SlugMsg, opts ...grpc.CallOption) (*PromoMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PromoMsg)
-	err := c.cc.Invoke(ctx, Promotion_GetPromotion_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Promotion_GetPromotion_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1059,8 +1118,9 @@ func (c *promotionClient) GetPromotion(ctx context.Context, in *SlugMsg, opts ..
 }
 
 func (c *promotionClient) UpdatePromotion(ctx context.Context, in *PromoWithSlug, opts ...grpc.CallOption) (*PromoMsg, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PromoMsg)
-	err := c.cc.Invoke(ctx, Promotion_UpdatePromotion_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Promotion_UpdatePromotion_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1068,8 +1128,9 @@ func (c *promotionClient) UpdatePromotion(ctx context.Context, in *PromoWithSlug
 }
 
 func (c *promotionClient) DeletePromotion(ctx context.Context, in *SlugMsg, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, Promotion_DeletePromotion_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Promotion_DeletePromotion_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1077,8 +1138,9 @@ func (c *promotionClient) DeletePromotion(ctx context.Context, in *SlugMsg, opts
 }
 
 func (c *promotionClient) ListPromotionItems(ctx context.Context, in *ListPromotionItemsReq, opts ...grpc.CallOption) (*PaginatedPromoItemsRes, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PaginatedPromoItemsRes)
-	err := c.cc.Invoke(ctx, Promotion_ListPromotionItems_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Promotion_ListPromotionItems_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1087,7 +1149,7 @@ func (c *promotionClient) ListPromotionItems(ctx context.Context, in *ListPromot
 
 // PromotionServer is the server API for Promotion service.
 // All implementations must embed UnimplementedPromotionServer
-// for forward compatibility
+// for forward compatibility.
 type PromotionServer interface {
 	ListPromotions(context.Context, *ListReq) (*PaginatedPromoRes, error)
 	PromotionSearch(context.Context, *SearchReq) (*PaginatedPromoRes, error)
@@ -1099,9 +1161,12 @@ type PromotionServer interface {
 	mustEmbedUnimplementedPromotionServer()
 }
 
-// UnimplementedPromotionServer must be embedded to have forward compatible implementations.
-type UnimplementedPromotionServer struct {
-}
+// UnimplementedPromotionServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPromotionServer struct{}
 
 func (UnimplementedPromotionServer) ListPromotions(context.Context, *ListReq) (*PaginatedPromoRes, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListPromotions not implemented")
@@ -1125,6 +1190,7 @@ func (UnimplementedPromotionServer) ListPromotionItems(context.Context, *ListPro
 	return nil, status.Errorf(codes.Unimplemented, "method ListPromotionItems not implemented")
 }
 func (UnimplementedPromotionServer) mustEmbedUnimplementedPromotionServer() {}
+func (UnimplementedPromotionServer) testEmbeddedByValue()                   {}
 
 // UnsafePromotionServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PromotionServer will
@@ -1134,6 +1200,13 @@ type UnsafePromotionServer interface {
 }
 
 func RegisterPromotionServer(s grpc.ServiceRegistrar, srv PromotionServer) {
+	// If the following call pancis, it indicates UnimplementedPromotionServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Promotion_ServiceDesc, srv)
 }
 
