@@ -3,10 +3,10 @@ package db
 import (
 	"context"
 	"errors"
-	repo "github.com/JMURv/par-pro/internal/repository"
-	"github.com/JMURv/par-pro/pkg/model"
-	gormutil "github.com/JMURv/par-pro/pkg/utils/gorm"
-	utils "github.com/JMURv/par-pro/pkg/utils/http"
+	repo "github.com/JMURv/par-pro/products/internal/repo"
+	"github.com/JMURv/par-pro/products/pkg/model"
+	gormutil "github.com/JMURv/par-pro/products/pkg/utils/gorm"
+	utils "github.com/JMURv/par-pro/products/pkg/utils/http"
 	"github.com/google/uuid"
 	"github.com/opentracing/opentracing-go"
 	"go.uber.org/zap"
